[Refector] 결제 서비스 리팩토링
* 결제 관련한 엔티티, dto의 가격 관련된 필드들 타입을 int에서 BigDecimal로 변경

+++ src/main/java/kr/hhplus/be/server/domain/payment/Payment.java
@@ -3,11 +3,12 @@ package kr.hhplus.be.server.domain.payment;
 import jakarta.persistence.*;
 import kr.hhplus.be.server.domain.common.BaseEntity;
 import kr.hhplus.be.server.domain.order.entity.Order;
-import kr.hhplus.be.server.domain.product.ProductStatus;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -25,10 +26,10 @@ public class Payment extends BaseEntity {
     private PaymentStatus status;
 
     @Column(name = "total_amt")
-    private int totalAmt;
+    private BigDecimal totalAmt;
 
     @Builder
-    public Payment(Order order, PaymentStatus status, int totalAmt) {
+    public Payment(Order order, PaymentStatus status, BigDecimal totalAmt) {
         this.order = order;
         this.status = status;
         this.totalAmt = totalAmt;
